Split ROS node and class into seperate files
